codec/packet: refactoring to allow zero-copy packet creation
